Extract PASV parsing into an FSM
We almost defintely won't use this in multiple
places but I think this logic fits better in the
FSM layer because that's (hopefully) where all
the response parsing will happen. Now the
logic in ftp::Client can focus on the high-level
operations of: setting transfer mode, getting a
passive connection, and opening a data socket.