Create FSM for directory commands, and use it
Directory commands == pwd, mkd.

These commands both return a path on success, which
needs to be parsed in the same way.

mkd doesn't currently return the path contained
in its response. Since they are sharing logic now,
change it so it does return the path, and update
a relevant test.